Types: re-locate the API specific types (#2096)
Specific commands have optimize result types, especially Geo and Streams for example - let's move all those out of the main file list to make maintenance easier.
