+++ src/StackExchange.Redis/APITypes/GeoSearchShape.cs
