+++ src/StackExchange.Redis/Enums/SimulatedFailureType.cs
